refactor: Update max-width values in News and NewsContainer styles
The max-width values in the News and NewsContainer styles were updated to improve the layout and accommodate larger screen sizes. The max-width in News.module.scss was changed from 1200px to 1400px, and the max-width in NewsContainer.module.scss was changed from 1180px to 1400px. These changes ensure that the content is displayed properly and optimally on a wider range of devices and screen resolutions.