refactor: Update About page to format and display company profile address
This commit refactors the About page to format and display the company profile address. It introduces a new variable, `profileAddress`, which uses the `formatTextWithLineBreaks` function to format the address with line breaks. The address is then displayed in the `dd` element using the `dangerouslySetInnerHTML` prop. This change improves the readability and presentation of the company's location information on the About page.