refactor: Enable lazy loading for images in Footer component
This commit enables lazy loading for the images in the Footer component. By adding the scroll={true} prop to the Link components, we optimize the loading and rendering of images based on the device screen size. This change improves page load performance and enhances the user experience by reducing the initial load time and optimizing the rendering of images on the page.